third_party/blink/web_tests/fast/events/pointerevents/device-id/device-properties-uniqueid-use-count.html

<!DOCTYPE HTML>
<script src='../../../../resources/testharness.js'></script>
<script src='../../../../resources/testharnessreport.js'></script>
<script src='../../../../resources/gesture-util.js'></script>

<style>
div.box {
  margin: 5px;
  padding: 20px;
  float: left;
  width: 50px;
  height: 50px;
}
</style>

<body>

<div id="grey" class="box" style="background-color:grey">
</div>

</body>

<script>
if (window.internals)
  internals.runtimeFlags.pointerEventDeviceIdEnabled = true;

var DevicePropertiesUniqueIdAttributeCount = 4943; // Comes from enum Feature in UseCounter.h

promise_test(async () => {
  var targetDiv = document.getElementById("grey");
  var uid = 0;
  var rect = document.getElementById("grey").getBoundingClientRect();
  var x1 = rect.left + 5;
  var y1 = rect.top + 5;

  targetDiv.addEventListener("pointerdown", function(event) {
    // Read uniqueId to trigger Measure.
    if (event.deviceProperties)
      uid = event.deviceProperties.uniqueId;
  });

  assert_false(internals.isUseCounted(document, DevicePropertiesUniqueIdAttributeCount));
  await mouseClickOn(x1, y1);
  await waitUntil( () => { return uid=-1; } );
  if (internals.runtimeFlags.pointerEventDeviceIdEnabled)
    assert_true(internals.isUseCounted(document, DevicePropertiesUniqueIdAttributeCount));
  else
    assert_false(internals.isUseCounted(document, DevicePropertiesUniqueIdAttributeCount));

}, "Verifies that use of deviceProperties.uniqueId is measured by use counters.");

</script>
